Add num requests to request_rate_sweep (#55)
* Same arrival times as vLLM

* Removed chunked prefill flags

* Fixed stray comment about chunked prefill

* Working simulator

* Fixed metrics calculation

* Initial chunked prefill

* Running chunked prefill

* Fixed stray comments

* Fixed conflicts in kvcache

* Running chunked prefill

* Added per-sec throughput calculation

* Added sweep params for chunked prefill and rates

* Added num requests to request_rate_sweep

* Small file format naming changes

* Fixed minor formatting issues

* Converted vllm estimated time to float

* Sweep over num_requests

* Lower bound TPOT to 0

* Modified ttft calculation

* Resolved repeated arrivaldeltas file creation

+++ sim/metrics.go
@@ -41,7 +41,7 @@ func (m *Metrics) Print(horizon int64, totalBlocks int, startTime time.Time) {
 	fmt.Printf("Total Input Tokens   : %d\n", m.TotalInputTokens)
 	fmt.Printf("Total Output Tokens  : %d\n", m.TotalOutputTokens)
 	fmt.Printf("Simulation Duration(s): %.3f\n", time.Since(startTime).Seconds())
-	fmt.Printf("vLLM estimated Duration(s): %d\n", m.SimEndedTime/1e6)
+	fmt.Printf("vLLM estimated Duration(s): %.3f\n", float64(m.SimEndedTime)/float64(1e6))
 	if m.CompletedRequests > 0 {
 		avgTTFT := float64(m.TTFTSum) / float64(m.CompletedRequests)
 		sortedTTFTs := SortRequestMetrics(m.RequestTTFTs)
